refactor: 모든 svg 파일 디렉토리별로 구분
