+++ src/pages/Auth/Signup/SignupPage/components/DomainDropdown.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import cn from 'utils/ts/classNames';
 import { useFormContext } from 'react-hook-form';
-import { ReactComponent as ArrowIcon } from 'assets/svg/arrow.svg';
+import { ReactComponent as ArrowIcon } from 'assets/svg/common/arrow.svg';
 import styles from '../SignUp.module.scss';
 import { DOMAIN, ERROR_MESSAGE } from '../../static/signUp';
 import { SignUpFormData } from '../entity';
